requirements.txt
protobuf==5.27.2
grpcio
grpcio-tools
cryptography
python-dotenv